Added: inputLimiter.js
Made it so that you can send in files as an input to be handled and displayed.
Noticed a problem so I scrapped that shit, because it said "fuck you, I ain't a file, I'm a string".
So I have just commented that this shit doesn't work + you can't select it, the function
is functioning.
Gonna focus more on the media-link way, of sending in an image or video and it will get embedded
and displayed appropriately.