Added: createATideHandler.js create-a-tide.sawcon
Modified:
userChecker.js -> alerts.js

What I have done now I added a new endpoint for adding tides, in addition I made it safer for the website
with taking into concern of dangerous characters.
I created 2 new methods for createATideHandler.js, where if there's multiple of same value in array, do not
proceed by putting the user to an alert.
The other one is to analyze the input for dangerous characters, applicaple for username and tides. If
dangerous characters are found, return true and then proceed by putting the user to an alert.
(I'm fucking proud of that shit)